moving out post_string to a util function
