+++ lib/ajax.js
@@ -1,4 +1,4 @@
-import { extend, generate_topic } from './util.js';
+import { extend, generate_topic, post_string } from './util.js';
 import setImmediate from './shims/setImmediate.js';
 import createJSONP from './internal/jsonp.js';
 
@@ -16,40 +16,6 @@ import createJSONP from './internal/jsonp.js';
 
 var _before = [],
 	_mocks = {};
-
-// processes an object into a querystring that can be passed to the server
-function postString(obj, prefix) {
-	var str = [], p, k, v;
-
-	if (Object.prototype.toString.call(obj) === '[object Array]') {
-		if (!prefix) {
-			throw new Error(
-				'Sorry buddy, your object is wrong' +
-				'and you should feel bad');
-		}
-
-		for (p = 0; p < obj.length; p++) {
-			k = prefix + '[]';
-			v = obj[p];
-			str.push(typeof v === 'object' ? postString(v, k) :
-				encodeURIComponent(k) + '=' + encodeURIComponent(v));
-		}
-	} else {
-		for (p in obj) {
-			if (prefix) {
-				k = prefix + '[' + p + ']';
-			} else {
-				k = p;
-			}
-
-			v = obj[p];
-			str.push(typeof v === 'object' ? postString(v, k) :
-				encodeURIComponent(k) + '=' + encodeURIComponent(v));
-		}
-	}
-
-	return str.join('&');
-}
 
 // slices variables out of the url and joins it with the data sent
 // to the ajax call
@@ -221,7 +187,7 @@ function Ajax(params) {
 	/* istanbul ignore else: POST will do its own thing */
 	if (_params.method === 'GET') {
 		_params.url += (_params.url.indexOf('?') === -1 ? '?' : '&') +
-			postString(_params.data);
+			post_string(_params.data);
 	}
 
 	_xhr.open(_params.method, _params.url, _params.async);
@@ -243,7 +209,7 @@ function Ajax(params) {
 		_before[ni](_xhr);
 	}
 
-	_xhr.send(_params.method === 'POST' ? postString(_params.data) : null);
+	_xhr.send(_params.method === 'POST' ? post_string(_params.data) : null);
 
 	return _xhr;
 }
