+++ lib/util.js
@@ -165,3 +165,39 @@ export function generate_topic(topic) {
 
 	return ret;
 }
+
+// processes an object into a querystring that can be passed to the server
+export function post_string(obj, prefix) {
+	var str = [], p, k, v;
+
+	if (Object.prototype.toString.call(obj) === '[object Array]') {
+		if (!prefix) {
+			throw new Error(
+				'Sorry buddy, your object is wrong' +
+				'and you should feel bad');
+		}
+
+		for (p = 0; p < obj.length; p++) {
+			k = prefix + '[]';
+			v = obj[p];
+			str.push(typeof v === 'object' ? post_string(v, k) :
+				encodeURIComponent(k) + '=' + encodeURIComponent(v));
+		}
+	} else {
+		for (p in obj) {
+			if (prefix) {
+				k = prefix + '[' + p + ']';
+			} else {
+				k = p;
+			}
+
+			v = obj[p];
+			str.push(typeof v === 'object' ? post_string(v, k) :
+				encodeURIComponent(k) + '=' + encodeURIComponent(v));
+		}
+	}
+
+	return str.join('&');
+}
+
+
